MWPW-147192 - formatDate moved to utils; NewsCard class moved to separate file;
